portmaster/requirements.txt
PyQt6>=6.6.0
psutil>=5.9.0
pyyaml>=6.0
toml>=0.10.2
